Added the script
Here is the carbot script.  Unfortunately this is an early version that just moves forward and stops when the ultrasonic sensor sees something within 30 inches.  I lost the final version shown in the video somehow.  If someone really wants to use the final version it should be trivial to add the loop, but you'll need to set new GPIO pins for the turning motor and calibrate that.